Fixed issue where addresses were uint_8s instead of uint_32s

+++ lib/WatchpointManager.cpp
@@ -4,15 +4,15 @@
 
 using namespace axe;
 
-void WatchpointManager::setWatchpoint(WatchpointType type, uint8_t lowAddr, uint8_t highAddr)
+void WatchpointManager::setWatchpoint(WatchpointType type, uint32_t lowAddr, uint32_t highAddr)
 {
-	Watchpoint w = Watchpoint(type, std::pair<uint8_t, uint8_t>(lowAddr, highAddr));
+	Watchpoint w = Watchpoint(type, std::pair<uint32_t, uint32_t>(lowAddr, highAddr));
 	watchpoints.insert(w);
 }
 
-void WatchpointManager::unsetWatchpoint(WatchpointType type, uint8_t lowAddr, uint8_t highAddr)
+void WatchpointManager::unsetWatchpoint(WatchpointType type, uint32_t lowAddr, uint32_t highAddr)
 {
-	Watchpoint w = Watchpoint(type, std::pair<uint8_t, uint8_t>(lowAddr, highAddr));
+	Watchpoint w = Watchpoint(type, std::pair<uint32_t, uint32_t>(lowAddr, highAddr));
 }
 
 bool WatchpointManager::contains(Watchpoint w)
